Prevent tutorial tooltip from appearing on every page.
Also, moved code of onboarding notifications from interactions.js to a
separate file.
